requirements.txt
apt-clone==0.2.1
apturl==0.5.2
async-timeout==4.0.1
attrs==21.2.0
bcrypt==5.0.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
eyeD3==0.8.10
filelock==3.6.0
greenlet==3.3.0
gyp==0.1
httplib2==0.20.2
hupper==1.12.1
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==6.0.2
lxml_html_clean==0.4.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
mutagen==1.45.1
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
PasteDeploy==3.1.0
pexpect==4.8.0
Pillow==9.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyramid==2.0.2
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.46
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
translationstring==1.4
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
venusian==3.1.1
wadllib==1.3.6
waitress==3.0.2
webencodings==0.5.1
WebOb==1.8.9
websocket-client==1.2.3
websockets==9.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.4.8
zipp==1.0.0
zope.deprecation==6.0
zope.interface==8.2
